feat(attendance): add attendance profiles, shift timings, and dialogs
