+++ features/attendance/index.ts
@@ -1,1 +1,6 @@
 export { MarkAttendanceForm } from './components/mark-attendance-form';
+export { ShiftTimings } from './components/shift-timings';
+export { ShiftDialog } from './components/shift-dialog';
+export { AttendanceProfiles } from './components/attendance-profiles';
+export { ProfileDialog } from './components/profile-dialog';
+export { DeleteConfirmDialog } from './components/delete-confirm-dialog';
